requirements.txt
fastapi[all]
pydantic
sqlalchemy
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp
pytest
httpx